% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraints.R
\name{AdministrationConstraints}
\alias{AdministrationConstraints}
\title{AdministrationConstraints}
\usage{
AdministrationConstraints(outcome = character(0), doses = list())
}
\arguments{
\item{outcome}{Character string: outcome name.}

\item{doses}{List of admissible dose levels for that outcome.}
}
\description{
Allowed dose values per outcome when optimizing or reporting designs.
}
