% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R, R/pfim-rd-examples.R
\name{armAdministration}
\alias{armAdministration}
\title{Tabular administration settings for reports}
\usage{
armAdministration(arm, designName, ...)
}
\arguments{
\item{arm}{First argument of generic.}

\item{...}{Not used.}
}
\value{
List of administration rows formatted for reports.
}
\description{
Tabular administration settings for reports

Default method for \code{Arm}.
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(armAdministration)
}
}
\keyword{internal}
