% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateTest.R
\name{CovariateTest-class}
\alias{CovariateTest-class}
\alias{CovariateTest}
\title{CovariateTest Class}
\arguments{
\item{significance}{A \code{data.frame} for the significance test results.}

\item{nonRelevance}{A \code{data.frame} for the non-relevance test results.}

\item{relevance}{A \code{data.frame} for the relevance test results.}
}
\description{
\code{CovariateTest} is the S7 class storing the covariate test results.
}
\details{
Three slots are available:
\describe{
  \item{significance}{Significance test (two-sided Wald, H0: beta = 0)}
  \item{nonRelevance}{Non-relevance / TOST test (H1: ratio in [0.80, 1.25])}
  \item{relevance}{Relevance test (H1: ratio outside [0.80, 1.25])}
}
}
