% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{Dcriterion}
\alias{Dcriterion}
\title{D-criterion}
\value{
Numeric D-optimality criterion value.
}
\description{
D-criterion
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
PFIM:::Dcriterion(prop(ev, "fim"))
}
}
\keyword{internal}
