% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\name{defineModelEquationsFromLibraryOfModel}
\alias{defineModelEquationsFromLibraryOfModel}
\title{Load model equations from the PK or PD library}
\description{
Load model equations from the PK or PD library
}
