% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{defineModelWrapper}
\alias{defineModelWrapper}
\title{Compile model equations and output mapping before evaluation}
\usage{
defineModelWrapper(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{\code{evaluation} or \code{arm}, depending on the method.}
}
\description{
Compile model equations and output mapping before evaluation
}
