% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Distribution}
\alias{Distribution}
\title{Distribution}
\usage{
Distribution(name = character(0), mu = 0, omega = 0)
}
\arguments{
\item{name}{Character string: distribution label.}

\item{mu}{Mean on the natural scale (or log-scale mean for log-normal).}

\item{omega}{Standard deviation of random effects (IIV).}
}
\description{
Base class for parameter distributions (\code{Normal}, \code{LogNormal}).
}
