% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{.assembleDesignFim}
\alias{.assembleDesignFim}
\title{Sum arm-level FIM matrices into a new design-level \code{Fim} object.}
\usage{
.assembleDesignFim(fimPrototype, evaluationArms)
}
\arguments{
\item{fimPrototype}{Template from \code{defineFim()} (class only; not mutated).}

\item{evaluationArms}{List of evaluated \code{Arm} objects.}
}
\value{
Aggregated \code{Fim} for the design.
}
\description{
Sum arm-level FIM matrices into a new design-level \code{Fim} object.
}
\keyword{internal}
