% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-fim.R
\name{.betaDictFromCovariates}
\alias{.betaDictFromCovariates}
\title{numeric vector of covariate effect sizes}
\usage{
.betaDictFromCovariates(modelCovariates)
}
\arguments{
\item{modelCovariates}{List of covariate objects.}
}
\value{
Named numeric vector.
}
\description{
numeric vector of covariate effect sizes
}
\keyword{internal}
