% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-fim.R
\name{.betaInternalNamesFromCovariates}
\alias{.betaInternalNamesFromCovariates}
\title{Internal beta column names (beta_param_covariate_category) from covariate definitions.}
\usage{
.betaInternalNamesFromCovariates(modelCovariates)
}
\arguments{
\item{modelCovariates}{List of covariate objects.}
}
\value{
Character vector.
}
\description{
Internal beta column names (beta_param_covariate_category) from covariate definitions.
}
\keyword{internal}
