% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{.constraintsKbl}
\alias{.constraintsKbl}
\title{.constraintsKbl - build a kableExtra table from arm constraints.}
\usage{
.constraintsKbl(arms, optimizationAlgorithm, col_names)
}
\arguments{
\item{arms}{List-of-arms}

\item{optimizationAlgorithm}{An Optimization subclass object.}

\item{col_names}{Character vector of column names for the table.}
}
\description{
.constraintsKbl - build a kableExtra table from arm constraints.
}
\keyword{internal}
