% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-fim-variance.R
\name{.evaluateVarianceFIMPop}
\alias{.evaluateVarianceFIMPop}
\title{Dispatch population variance FIM between simple and covariate/occasion paths.}
\usage{
.evaluateVarianceFIMPop(fim, model, arm, isFixedMu = NULL, isFixedOmega = NULL)
}
\description{
Dispatch population variance FIM between simple and covariate/occasion paths.
}
\keyword{internal}
