% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-fim-variance.R
\name{.evaluateVarianceFIMPopCovariateOccasion}
\alias{.evaluateVarianceFIMPopCovariateOccasion}
\title{Covariate / occasion path: sum Fisher blocks over covariate combinations.}
\usage{
.evaluateVarianceFIMPopCovariateOccasion(model, arm, isFixedMu, isFixedOmega)
}
\description{
Covariate / occasion path: sum Fisher blocks over covariate combinations.
}
\keyword{internal}
