% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-fim-variance.R
\name{.evaluateVarianceFIMPopSimple}
\alias{.evaluateVarianceFIMPopSimple}
\title{Standard population path: fixed mu, diagonal IIV, no occasion structure.}
\usage{
.evaluateVarianceFIMPopSimple(model, arm, isFixedMu, isFixedOmega)
}
\description{
Standard population path: fixed mu, diagonal IIV, no occasion structure.
}
\keyword{internal}
