% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-project-access.R
\name{.evaluationFromOptimization}
\alias{.evaluationFromOptimization}
\title{Build an \code{Evaluation} from an \code{Optimization} and a single \code{Design}.}
\usage{
.evaluationFromOptimization(optimization, design, name = NULL)
}
\description{
Build an \code{Evaluation} from an \code{Optimization} and a single \code{Design}.
}
\keyword{internal}
