% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{.getListLastName}
\alias{.getListLastName}
\title{Names of the deepest elements in a nested list (used for model-type detection).}
\usage{
.getListLastName(list)
}
\description{
Names of the deepest elements in a nested list (used for model-type detection).
}
\keyword{internal}
