% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\docType{data}
\name{.greekConsole}
\alias{.greekConsole}
\title{Console / plain-text display (Unicode, separator already included)}
\format{
An object of class \code{character} of length 5.
}
\usage{
.greekConsole
}
\description{
Console / plain-text display (Unicode, separator already included)
}
\keyword{datasets}
