% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\docType{data}
\name{.greekLatex}
\alias{.greekLatex}
\title{LaTeX for kableExtra HTML tables}
\format{
An object of class \code{character} of length 5.
}
\usage{
.greekLatex
}
\description{
LaTeX for kableExtra HTML tables
}
\keyword{datasets}
