% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\docType{data}
\name{.greekPlot}
\alias{.greekPlot}
\title{ggplot axis labels (bare Unicode, no separator)}
\format{
An object of class \code{character} of length 4.
}
\usage{
.greekPlot
}
\description{
ggplot axis labels (bare Unicode, no separator)
}
\keyword{datasets}
