% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-fim.R
\name{.isNestedArmEvaluation}
\alias{.isNestedArmEvaluation}
\title{isNestedArmEvaluation}
\usage{
.isNestedArmEvaluation(x)
}
\arguments{
\item{x}{\code{evaluationGradients} or \code{evaluationVariance} from an arm.}
}
\value{
Logical scalar.
}
\description{
isNestedArmEvaluation
}
\keyword{internal}
