% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{.pfimConstraintTaskIndices}
\alias{.pfimConstraintTaskIndices}
\title{Stratified subsample of constraint-grid cells}
\usage{
.pfimConstraintTaskIndices(totalIterations, numberOfDoses, nCombinations)
}
\description{
Stratified subsample of constraint-grid cells
}
\keyword{internal}
