% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-fim-cache.R
\name{.pfimEvaluationFromDesign}
\alias{.pfimEvaluationFromDesign}
\title{Copy FIM results from a finished evaluationDesign back into an Evaluation shell.}
\usage{
.pfimEvaluationFromDesign(evaluation, design, evaluatedDesign)
}
\description{
Copy FIM results from a finished evaluationDesign back into an Evaluation shell.
}
\keyword{internal}
