% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-fim-cache.R
\name{.pfimRunEvaluations}
\alias{.pfimRunEvaluations}
\title{Run multiple evaluations through the FIM cache.}
\usage{
.pfimRunEvaluations(evaluations)
}
\description{
Run multiple evaluations through the FIM cache.
}
\keyword{internal}
