% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-fim-cache.R
\name{.pfimRunOptimizationEvaluation}
\alias{.pfimRunOptimizationEvaluation}
\title{Run a final optimization design evaluation through the FIM cache.}
\usage{
.pfimRunOptimizationEvaluation(optimization, design, name = "")
}
\description{
Run a final optimization design evaluation through the FIM cache.
}
\keyword{internal}
