% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{.renderEvalReport}
\alias{.renderEvalReport}
\title{create a \code{generateReportEvaluation} method body.}
\usage{
.renderEvalReport(template)
}
\arguments{
\item{template}{Basename of the evaluation Rmd template.}
}
\value{
A function with the standard \code{generateReportEvaluation} signature.
}
\description{
create a \code{generateReportEvaluation} method body.
}
\keyword{internal}
