% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{.reportTemplatePath}
\alias{.reportTemplatePath}
\title{Absolute path to an Rmd skeleton template in the PFIM package.
Called lazily at render time, never at package load time.}
\usage{
.reportTemplatePath(filename)
}
\arguments{
\item{filename}{Basename of the Rmd template file.}
}
\description{
Absolute path to an Rmd skeleton template in the PFIM package.
Called lazily at render time, never at package load time.
}
\keyword{internal}
