% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-utils.R
\name{ensureModelOutputNames}
\alias{ensureModelOutputNames}
\title{Ensure \code{outputNames} is filled from project \code{outputs} when empty.}
\usage{
ensureModelOutputNames(model, pfimproject)
}
\description{
Ensure \code{outputNames} is filled from project \code{outputs} when empty.
}
\keyword{internal}
