% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalyticInfusion.R
\name{evaluateAnalyticInfusionCore}
\alias{evaluateAnalyticInfusionCore}
\title{Evaluate analytic infusion predictions (superposition of during/after doses).}
\usage{
evaluateAnalyticInfusionCore(model, arm)
}
\arguments{
\item{model}{A \code{ModelAnalyticInfusion} object.}

\item{arm}{An \code{Arm} object.}
}
\value{
Named list of output data frames at sampling times.
}
\description{
Unlike the steady-state infusion core, arguments are passed as named lists
(vectorised over past doses when several infusions contribute).
}
\keyword{internal}
