% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R, R/pfim-rd-examples.R
\name{evaluateErrorModelDerivatives}
\alias{evaluateErrorModelDerivatives}
\title{Evaluate residual variance and dV/dsigma diagonals}
\usage{
evaluateErrorModelDerivatives(modelError, ...)
}
\arguments{
\item{modelError}{A \code{ModelError} object.}

\item{...}{For \code{ModelError}: \code{evaluationModel} (numeric predictions).}
}
\value{
List with \code{errorVariance} and \code{sigmaDerivatives}.
}
\description{
Evaluate residual variance and dV/dsigma diagonals
}
\examples{
err = Combined2(output = "RespPK", sigmaInter = 0.2, sigmaSlope = 0.1)
PFIM:::evaluateErrorModelDerivatives(err, c(1, 2))
}
\keyword{internal}
