% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{evaluateFim}
\alias{evaluateFim}
\title{Compute the FIM for one arm}
\description{
Compute the FIM for one arm

Compute the FIM for one arm

Compute the FIM for one arm
}
