% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Predict model responses at arm sampling times}
\usage{
evaluateModel(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{\code{arm} for model methods.}
}
\description{
Predict model responses at arm sampling times
}
