% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateModelGradient}
\alias{evaluateModelGradient}
\title{Gradients of model responses with respect to parameters}
\usage{
evaluateModelGradient(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{\code{arm} for model methods.}
}
\description{
Gradients of model responses with respect to parameters
}
