% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-gradient.R
\name{evaluateModelGradientCore}
\alias{evaluateModelGradientCore}
\title{Evaluate gradient core without covariate expansion.}
\usage{
evaluateModelGradientCore(model, arm)
}
\arguments{
\item{model}{\code{Model} configured for gradient computation.}

\item{arm}{\code{Arm} object used for model evaluation.}
}
\value{
Named list of gradient data frames by output.
}
\description{
Evaluate gradient core without covariate expansion.
}
\keyword{internal}
