% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-gradient.R
\name{evaluateModelGradientWithCovariates}
\alias{evaluateModelGradientWithCovariates}
\title{Evaluate gradients with covariate and occasion structure.}
\usage{
evaluateModelGradientWithCovariates(model, arm, evaluateModelGradientCore)
}
\arguments{
\item{model}{\code{Model} containing covariate combinations.}

\item{arm}{\code{Arm} object used for evaluation.}

\item{evaluateModelGradientCore}{Function evaluating gradients for one occasion model.}
}
\value{
Nested list of gradients by combination and occasion.
}
\description{
Evaluate gradients with covariate and occasion structure.
}
\keyword{internal}
