% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateModelVariance}
\alias{evaluateModelVariance}
\title{Residual variance at sampling times for an arm}
\usage{
evaluateModelVariance(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{Arm passed to methods.}
}
\description{
Residual variance at sampling times for an arm
}
\keyword{internal}
