% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/model-covariates.R
\name{evaluateModelWithCovariates}
\alias{evaluateModelWithCovariates}
\title{Evaluate model outputs over covariate combinations and occasions}
\usage{
evaluateModelWithCovariates(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{Arm and evaluation core function (see methods).}
}
\value{
Nested list of evaluations keyed by combination and occasion.
}
\description{
Evaluate model outputs over covariate combinations and occasions

Evaluate model outputs over every covariate combination and occasion.
}
\keyword{internal}
