% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{FedorovWynnAlgorithm}
\alias{FedorovWynnAlgorithm}
\title{FedorovWynnAlgorithm}
\usage{
FedorovWynnAlgorithm(
  elementaryProtocols = list(),
  numberOfSubjects = 0,
  proportionsOfSubjects = 0,
  showProcess = FALSE,
  FedorovWynnAlgorithmOutputs = list()
)
}
\arguments{
\item{elementaryProtocols}{List of elementary protocols.}

\item{numberOfSubjects}{Total number of subjects.}

\item{proportionsOfSubjects}{Initial proportion vector for each protocol.}

\item{showProcess}{Logical; print FIM enumeration progress.}

\item{FedorovWynnAlgorithmOutputs}{Output list from the Rcpp routine.}
}
\description{
Fedorov-Wynn exchange algorithm for D-optimal sampling-time design (Rcpp kernel).
}
\details{
FedorovWynnAlgorithm: S7 class
}
