% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{finiteDifferenceHessian}
\alias{finiteDifferenceHessian}
\title{Finite-difference perturbations for gradient computation}
\usage{
finiteDifferenceHessian(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{Optional method arguments.}
}
\description{
Finite-difference perturbations for gradient computation
}
