% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{fun_amoeba}
\alias{fun_amoeba}
\title{Nelder-Mead simplex minimiser (R interface).}
\usage{
fun_amoeba(p, y, ftol, itmax, funk, outcomes, data, showProcess)
}
\arguments{
\item{p}{parameter p}

\item{y}{parameter y}

\item{ftol}{parameter ftol}

\item{itmax}{parameter itmax}

\item{funk}{parameter funk}

\item{outcomes}{The model outcomes.}

\item{data}{parameter data}

\item{showProcess}{Boolean.}
}
\value{
List with simplex results from the compiled kernel.
}
\description{
Nelder-Mead simplex minimiser (R interface).
}
\keyword{internal}
