% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-arm-constraints.R, R/pfim-rd-examples.R
\name{getArmConstraints}
\alias{getArmConstraints}
\title{Administration and sampling constraints for an optimizer}
\usage{
getArmConstraints(arm, optimizationAlgorithm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{optimizationAlgorithm}{An optimization algorithm object.}

\item{...}{Not used.}
}
\value{
A list of arm-level optimization constraints.
}
\description{
Administration and sampling constraints for an optimizer
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getArmConstraints)
}
}
\keyword{internal}
