% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmData}
\alias{getArmData}
\title{Dose and sampling summary for reports}
\usage{
getArmData(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Not used.}
}
\description{
Dose and sampling summary for reports
}
