% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-fim.R
\name{getArmEvaluationVarianceFlat}
\alias{getArmEvaluationVarianceFlat}
\title{Flat residual variance for Individual/Bayesian FIM.}
\usage{
getArmEvaluationVarianceFlat(arm)
}
\arguments{
\item{arm}{An \code{Arm} object.}
}
\value{
List with \code{errorVariance} and \code{sigmaDerivatives}.
}
\description{
Flat residual variance for Individual/Bayesian FIM.
}
\keyword{internal}
