% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariate.R, R/pfim-rd-examples.R
\name{getCategoryOfReference}
\alias{getCategoryOfReference}
\title{Return the reference category for a covariate}
\usage{
getCategoryOfReference(covariate, ...)
}
\arguments{
\item{covariate}{A \code{Covariate} object.}

\item{...}{Optional method arguments.}
}
\value{
A character value naming the reference category.
}
\description{
Return the reference category for a covariate

Reference category for a covariate (first element of \code{categories}).
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getCategoryOfReference)
}
}
\keyword{internal}
