% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/evaluation-accessors.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Parameter correlations from the optimal design FIM}
\description{
Parameter correlations from the optimal design FIM

Parameter correlations from the optimal design FIM
}
