% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R, R/pfim-rd-examples.R
\name{getModelErrorData}
\alias{getModelErrorData}
\title{Extract residual error settings for reports}
\usage{
getModelErrorData(modelError, ...)
}
\arguments{
\item{modelError}{A \code{ModelError} object.}

\item{...}{Optional method arguments.}
}
\value{
One-row data frame of residual-error settings.
}
\description{
Extract residual error settings for reports
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getModelErrorData)
}
}
\keyword{internal}
