% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameter.R, R/pfim-rd-examples.R
\name{getModelParametersData}
\alias{getModelParametersData}
\title{Parameter table for reports (mu, omega2, gamma2, fix flags, distribution).}
\arguments{
\item{modelParameter}{A \code{ModelParameter} object.}
}
\value{
A data frame of model parameter values.
}
\description{
Parameter table for reports (mu, omega2, gamma2, fix flags, distribution).
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getModelParametersData)
}
}
