% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{getRealisedDcriterion}
\alias{getRealisedDcriterion}
\title{D-criterion of the single protocol implemented after Multiplicative allocation.}
\usage{
getRealisedDcriterion(pfimproject, ...)
}
\arguments{
\item{pfimproject}{An \code{Optimization} from \code{MultiplicativeAlgorithm}.}

\item{...}{Optional method arguments.}
}
\value{
Numeric D-criterion of the implemented protocol.
}
\description{
Matches \code{getDcriterion()} on the optimal evaluation. Requires
\code{MultiplicativeAlgorithm} results.
}
