% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/evaluation-accessors.R
\name{getRSE}
\alias{getRSE}
\title{Relative standard errors from the optimal design FIM}
\description{
Relative standard errors from the optimal design FIM

Relative standard errors from the optimal design FIM
}
