% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R, R/pfim-rd-examples.R
\name{getSamplingData}
\alias{getSamplingData}
\title{Sampling times and grid extent for evaluation plots}
\usage{
getSamplingData(arm, ...)
}
\arguments{
\item{arm}{First argument of generic.}

\item{...}{Not used.}
}
\value{
List containing per-outcome sampling objects and numeric grids.
}
\description{
Sampling times and grid extent for evaluation plots

Default method for \code{Arm}.
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getSamplingData)
}
}
\keyword{internal}
