% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/evaluation-accessors.R
\name{getShrinkage}
\alias{getShrinkage}
\title{Parameter shrinkage from the Bayesian FIM}
\description{
Parameter shrinkage from the Bayesian FIM

Parameter shrinkage from the Bayesian FIM
}
