% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R, R/model-covariates.R,
%   R/pfim-rd-examples.R
\name{hasCovariates}
\alias{hasCovariates}
\title{Report whether a model includes covariates}
\usage{
hasCovariates(model, ...)
}
\arguments{
\item{model}{First argument of generic.}

\item{...}{Optional method arguments.}
}
\value{
Logical scalar indicating whether covariates are configured.
}
\description{
Report whether a model includes covariates

Report whether the model defines any covariates.
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(hasCovariates)
}
}
\keyword{internal}
