% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{inferNumberOfOccasions}
\alias{inferNumberOfOccasions}
\title{Infer the number of study occasions for a model.}
\usage{
inferNumberOfOccasions(modelCovariates = list(), modelParameters = list())
}
\arguments{
\item{modelCovariates}{List of covariate objects.}

\item{modelParameters}{List of \code{ModelParameter} objects.}
}
\value{
Integer number of occasions.
}
\description{
Rules (in order):
\enumerate{
  \item If any \code{CategoricalCovariateWithIOV} is present, use the common
        sequence length (e.g. 2, 3, or 4 periods).
  \item Else if any parameter has \code{gamma > 0} (random IOV), use 2 occasions.
  \item Else use 1 occasion.
}
}
\keyword{internal}
