% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfModels.R
\name{LibraryOfModels}
\alias{LibraryOfModels}
\title{LibraryOfModels}
\usage{
LibraryOfModels(models = list())
}
\arguments{
\item{models}{List of PK and PD model metadata objects.}
}
\description{
Registry of built-in PK and PD model definitions.
}
