% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2BolusSingleDose_ClQV1V2}
\alias{Linear2BolusSingleDose_ClQV1V2}
\title{Analytic PK equation: Linear2BolusSingleDose_ClQV1V2.}
\usage{
Linear2BolusSingleDose_ClQV1V2()
}
\value{
Character string: closed-form PK equation.
}
\description{
Two-compartment IV bolus (Cl, Q, V1, V2 parameterization).
}
\examples{
nzchar(Linear2BolusSingleDose_ClQV1V2())
}
