% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2BolusSingleDose_kk12k21V}
\alias{Linear2BolusSingleDose_kk12k21V}
\title{Analytic PK equation: Linear2BolusSingleDose_kk12k21V (micro-constant form).}
\usage{
Linear2BolusSingleDose_kk12k21V()
}
\value{
Character string: closed-form PK equation.
}
\description{
Analytic PK equation: Linear2BolusSingleDose_kk12k21V (micro-constant form).
}
\examples{
nzchar(Linear2BolusSingleDose_kk12k21V())
}
