% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2BolusSteadyState_ClQV1V2tau}
\alias{Linear2BolusSteadyState_ClQV1V2tau}
\title{Analytic PK equation: Linear2BolusSteadyState_ClQV1V2tau.}
\usage{
Linear2BolusSteadyState_ClQV1V2tau()
}
\value{
Character string: closed-form PK equation.
}
\description{
Analytic PK equation: Linear2BolusSteadyState_ClQV1V2tau.
}
\examples{
nzchar(Linear2BolusSteadyState_ClQV1V2tau())
}
