% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2BolusSteadyState_kk12k21Vtau}
\alias{Linear2BolusSteadyState_kk12k21Vtau}
\title{Analytic PK equation: Linear2BolusSteadyState_kk12k21Vtau.}
\usage{
Linear2BolusSteadyState_kk12k21Vtau()
}
\value{
Character string: closed-form PK equation.
}
\description{
Analytic PK equation: Linear2BolusSteadyState_kk12k21Vtau.
}
\examples{
nzchar(Linear2BolusSteadyState_kk12k21Vtau())
}
