% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2FirstOrderSingleDose_kaClQV1V2}
\alias{Linear2FirstOrderSingleDose_kaClQV1V2}
\title{Analytic PK equation: Linear2FirstOrderSingleDose_kaClQV1V2.}
\usage{
Linear2FirstOrderSingleDose_kaClQV1V2()
}
\value{
Character string: closed-form PK equation.
}
\description{
Analytic PK equation: Linear2FirstOrderSingleDose_kaClQV1V2.
}
\examples{
nzchar(Linear2FirstOrderSingleDose_kaClQV1V2())
}
