% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2FirstOrderSingleDose_kakk12k21V}
\alias{Linear2FirstOrderSingleDose_kakk12k21V}
\title{Analytic PK equation: Linear2FirstOrderSingleDose_kakk12k21V.}
\usage{
Linear2FirstOrderSingleDose_kakk12k21V()
}
\value{
Character string: closed-form PK equation.
}
\description{
Analytic PK equation: Linear2FirstOrderSingleDose_kakk12k21V.
}
\examples{
nzchar(Linear2FirstOrderSingleDose_kakk12k21V())
}
