% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2InfusionSingleDose_ClQV1V2}
\alias{Linear2InfusionSingleDose_ClQV1V2}
\title{Analytic PK equation: Linear2InfusionSingleDose_ClQV1V2.}
\usage{
Linear2InfusionSingleDose_ClQV1V2()
}
\value{
Named list with \code{duringInfusion} / \code{afterInfusion} equation strings.
}
\description{
Analytic PK equation: Linear2InfusionSingleDose_ClQV1V2.
}
\examples{
nzchar(Linear2InfusionSingleDose_ClQV1V2()$duringInfusion$RespPK)
}
