% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2InfusionSingleDose_kk12k21V}
\alias{Linear2InfusionSingleDose_kk12k21V}
\title{Analytic PK equation: Linear2InfusionSingleDose_kk12k21V.}
\usage{
Linear2InfusionSingleDose_kk12k21V()
}
\value{
Named list with \code{duringInfusion} / \code{afterInfusion} equation strings.
}
\description{
Analytic PK equation: Linear2InfusionSingleDose_kk12k21V.
}
\examples{
nzchar(Linear2InfusionSingleDose_kk12k21V()$duringInfusion$RespPK)
}
