% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2InfusionSteadyState_ClQV1V2tau}
\alias{Linear2InfusionSteadyState_ClQV1V2tau}
\title{Analytic PK equation: Linear2InfusionSteadyState_ClQV1V2tau.}
\usage{
Linear2InfusionSteadyState_ClQV1V2tau()
}
\value{
Named list with \code{duringInfusion} / \code{afterInfusion} equation strings.
}
\description{
Analytic PK equation: Linear2InfusionSteadyState_ClQV1V2tau.
}
\examples{
nzchar(Linear2InfusionSteadyState_ClQV1V2tau()$duringInfusion$RespPK)
}
