% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R, R/pfim-rd-examples.R
\name{Linear2InfusionSteadyState_kk12k21Vtau}
\alias{Linear2InfusionSteadyState_kk12k21Vtau}
\title{Analytic PK equation: Linear2InfusionSteadyState_kk12k21Vtau.}
\usage{
Linear2InfusionSteadyState_kk12k21Vtau()
}
\value{
Named list with \code{duringInfusion} / \code{afterInfusion} equation strings.
}
\description{
Analytic PK equation: Linear2InfusionSteadyState_kk12k21Vtau.
}
\examples{
nzchar(Linear2InfusionSteadyState_kk12k21Vtau()$duringInfusion$RespPK)
}
