% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelError.R
\name{ModelError}
\alias{ModelError}
\title{ModelError}
\usage{
ModelError(
  output = "output",
  equation = expression(),
  derivatives = list(),
  sigmaInter = 0,
  sigmaSlope = 0,
  sigmaInterFixed = FALSE,
  sigmaSlopeFixed = FALSE,
  cError = 1
)
}
\arguments{
\item{output}{Character string: outcome name.}

\item{equation}{Expression defining residual variance as a function of predictions.}

\item{derivatives}{List of partial derivatives w.r.t. error parameters.}

\item{sigmaInter}{Additive error standard deviation (inter-subject / residual).}

\item{sigmaSlope}{Proportional error standard deviation (slope).}

\item{sigmaInterFixed}{If TRUE, \code{sigmaInter} is fixed in the FIM.}

\item{sigmaSlopeFixed}{If TRUE, \code{sigmaSlope} is fixed in the FIM.}

\item{cError}{Power parameter in the error model (when applicable).}
}
\description{
Base class for residual error models (\code{Constant}, \code{Combined1}, \code{Proportional}).
}
