% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{Normal}
\alias{Normal}
\title{Normal}
\usage{
Normal(name = character(0), mu = 0, omega = 0)
}
\arguments{
\item{name}{Character string: distribution label.}

\item{mu}{Mean on the natural scale (or log-scale mean for log-normal).}

\item{omega}{Standard deviation of random effects (IIV).}
}
\description{
Normal distribution for model parameters (identity link on the mean).
}
