% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{parseCombinationName}
\alias{parseCombinationName}
\title{Parse combination name and fill missing reference values.}
\usage{
parseCombinationName(combinationName, modelCovariates)
}
\arguments{
\item{combinationName}{Character scalar combination identifier.}

\item{modelCovariates}{List of covariate objects used to infer defaults.}
}
\value{
Named list of covariate values including references.
}
\description{
Parse combination name and fill missing reference values.
}
\keyword{internal}
