% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPDModels.R
\name{pdModelLibrary}
\alias{pdModelLibrary}
\title{Default built-in PD model library instance.}
\format{
An S7 object of class \code{LibraryOfPDModels}.
}
\usage{
pdModelLibrary
}
\value{
A \code{LibraryOfPDModels} object.
}
\description{
Default built-in PD model library instance.
}
\examples{
\donttest{
length(prop(pdModelLibrary, "models"))
}
}
\seealso{
\code{\link{LibraryOfPDModels}} for the class constructor.
}
