% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-fim-cache.R, R/pfim-rd-examples.R
\name{pfim_cache_stats}
\alias{pfim_cache_stats}
\title{FIM cache sizes and hit counts}
\usage{
pfim_cache_stats()
}
\value{
A list of cache statistics.
}
\description{
FIM cache sizes and hit counts
}
\examples{
PFIM:::pfim_cache_stats()
}
\keyword{internal}
