% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-extensions.R, R/pfim-rd-examples.R
\name{pfim_list_extensions}
\alias{pfim_list_extensions}
\title{List registered extensions}
\usage{
pfim_list_extensions()
}
\value{
A list of registered extension names.
}
\description{
List registered extensions
}
\examples{
pfim_list_extensions()
}
