% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-extensions.R
\name{pfim_registered_model_classes}
\alias{pfim_registered_model_classes}
\alias{pfim_registered_fim_types}
\alias{pfim_registered_optimizers}
\title{Registered model class names}
\usage{
pfim_registered_model_classes()

pfim_registered_fim_types()

pfim_registered_optimizers()
}
\value{
A character vector of registered class names.
}
\description{
Registered model class names
}
\examples{
\donttest{
pfim_registered_model_classes()
pfim_registered_fim_types()
pfim_registered_optimizers()
}
}
