% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-session.R
\name{pfim_set_option}
\alias{pfim_set_option}
\title{Set PFIM session options}
\usage{
pfim_set_option(...)
}
\arguments{
\item{...}{Named option values (\code{fim.cache = FALSE}, etc.).}
}
\value{
Invisibly \code{NULL}.
}
\description{
Set PFIM session options
}
