% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGBOAlgorithm.R
\name{PGBOAlgorithm}
\alias{PGBOAlgorithm}
\title{PGBOAlgorithm}
\usage{
PGBOAlgorithm(
  N = numeric(0),
  muteEffect = numeric(0),
  maxIteration = numeric(0),
  purgeIteration = numeric(0),
  fitBase = 0.03,
  seed = numeric(0),
  showProcess = FALSE
)
}
\arguments{
\item{N}{Numeric: the parameter N.}

\item{muteEffect}{Numeric: the parameter muteEffect.}

\item{maxIteration}{Numeric: the parameter maxIteration.}

\item{purgeIteration}{Numeric: the parameter purgeIteration.}

\item{fitBase}{Numeric: baseline fitness for PGBO (default 0.03).}

\item{seed}{Numeric: the parameter seed.}

\item{showProcess}{Logical: showProcess.}
}
\description{
Population-based global optimization for sampling times (Rcpp kernel).
}
