% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R
\docType{data}
\name{pkModelLibrary}
\alias{pkModelLibrary}
\title{Default built-in PK model library instance.}
\format{
An S7 object of class \code{LibraryOfPKModels}.
}
\usage{
pkModelLibrary
}
\description{
Default built-in PK model library instance.
}
\seealso{
\code{\link{LibraryOfPKModels}} for the class constructor.
}
\keyword{datasets}
