% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationResults}
\alias{plotEvaluationResults}
\title{ggplot of predicted responses with sampling points}
\usage{
plotEvaluationResults(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Method-specific plotting inputs (see methods).}
}
\value{
Nested list of ggplot objects for model responses.
}
\description{
ggplot of predicted responses with sampling points

Default method for \code{Arm}.
}
\keyword{internal}
