% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationSI}
\alias{plotEvaluationSI}
\title{ggplot of parameter sensitivities over time}
\usage{
plotEvaluationSI(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Gradient data and plot options for the \code{Arm} method.}
}
\description{
ggplot of parameter sensitivities over time
}
