% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/pfim-rd-examples.R
\name{plotFrequencies}
\alias{plotFrequencies}
\title{Plot optimized sampling frequencies}
\usage{
plotFrequencies(optimization, ...)
}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{...}{Optional method arguments.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plot optimized sampling frequencies

Fedorov-Wynn optimal frequencies
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(plotFrequencies)
}
}
