% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{plotFrequenciesFedorovWynnAlgorithm}
\alias{plotFrequenciesFedorovWynnAlgorithm}
\title{Frequency trajectories of the Fedorov-Wynn algorithm}
\description{
Frequency trajectories of the Fedorov-Wynn algorithm
}
