% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{plotRSEFIM}
\alias{plotRSEFIM}
\title{Plot relative standard errors from a FIM}
\usage{
plotRSEFIM(fim, evaluation, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{evaluation}{A \code{PFIMProject} evaluation object.}

\item{...}{Graphics parameters passed to \code{ggplot2}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plot relative standard errors from a FIM
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
PFIM:::plotRSEFIM(prop(ev, "fim"), ev)
}
}
\keyword{internal}
