% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{plotSEFIM}
\alias{plotSEFIM}
\title{Barplot of standard errors from a FIM}
\usage{
plotSEFIM(fim, evaluation, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{evaluation}{A \code{PFIMProject} object.}

\item{\dots}{Additional arguments passed to methods.}
}
\description{
Barplot of standard errors from a FIM

Barplot of standard errors from a FIM

Barplot of standard errors from a FIM

Barplot of standard errors from a FIM
}
