% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/pfim-rd-examples.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{Plot Bayesian shrinkage from a FIM}
\usage{
plotShrinkage(fim, evaluation, ...)
}
\arguments{
\item{fim}{First argument of generic.}

\item{evaluation}{\code{PFIMProject} providing model parameter labels.}

\item{...}{Graphics parameters passed to \code{ggplot2}.}
}
\value{
\code{ggplot} object showing Bayesian shrinkage by parameter.
}
\description{
Plot Bayesian shrinkage from a FIM

Default method for \code{BayesianFim}.
}
\examples{
\dontrun{
source(system.file("examples", "covariate-test-minimal.R", package = "PFIM"))
# plotShrinkage(prop(ev, "fim"), ev)  # BayesianFim
}
}
