% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/pfim-rd-examples.R
\name{plotWeights}
\alias{plotWeights}
\title{Plot optimized arm weights}
\usage{
plotWeights(optimization, ...)
}
\arguments{
\item{optimization}{An \code{Optimization} object.}

\item{...}{Optional method arguments.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plot optimized arm weights

Multiplicative algorithm weights by iteration
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(plotWeights)
}
}
