% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R
\name{plotWeightsMultiplicativeAlgorithm}
\alias{plotWeightsMultiplicativeAlgorithm}
\title{Weight trajectories of the multiplicative algorithm}
\description{
Weight trajectories of the multiplicative algorithm
}
