% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{processArmEvaluationResults}
\alias{processArmEvaluationResults}
\title{Model response plots for one arm}
\usage{
processArmEvaluationResults(arm, model, fim, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{model}{A \code{Model} object.}

\item{fim}{A \code{Fim} object.}

\item{...}{\code{designName} and \code{plotOptions}.}
}
\description{
Model response plots for one arm
}
