% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-library-remap.R
\name{remapOdePkLibraryEquations}
\alias{remapOdePkLibraryEquations}
\title{Remap an ODE PK equation list onto project compartments.}
\usage{
remapOdePkLibraryEquations(equations, pfimproject)
}
\arguments{
\item{equations}{Character vector or during/after infusion lists.}

\item{pfimproject}{A \code{PFIMProject} object.}
}
\value{
Remapped equations.
}
\description{
Remap an ODE PK equation list onto project compartments.
}
\keyword{internal}
