% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-library-remap.R
\name{remapOdePkLibraryText}
\alias{remapOdePkLibraryText}
\title{Remap one ODE PK library equation string (\code{C1}/\code{C2} tokens).}
\usage{
remapOdePkLibraryText(text, pfimproject)
}
\arguments{
\item{text}{Character equation string.}

\item{pfimproject}{A \code{PFIMProject} object.}
}
\value{
Updated character string.
}
\description{
Remap one ODE PK library equation string (\code{C1}/\code{C2} tokens).
}
\keyword{internal}
