% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-library-remap.R
\name{remapPkpdLibraryEquations}
\alias{remapPkpdLibraryEquations}
\title{Remap a PK/PD equation list onto project compartments.}
\usage{
remapPkpdLibraryEquations(equations, pfimproject)
}
\arguments{
\item{equations}{Character vector or during/after infusion lists.}

\item{pfimproject}{A \code{PFIMProject} object.}
}
\value{
Remapped equations.
}
\description{
Remap a PK/PD equation list onto project compartments.
}
\keyword{internal}
