% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-library-remap.R
\name{remapPkpdLibraryText}
\alias{remapPkpdLibraryText}
\title{Remap one PK/PD library equation string onto project compartments.}
\usage{
remapPkpdLibraryText(text, pfimproject)
}
\arguments{
\item{text}{Character equation string.}

\item{pfimproject}{A \code{PFIMProject} object.}
}
\value{
Updated character string.
}
\description{
Remap one PK/PD library equation string onto project compartments.
}
\keyword{internal}
