% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{resolveNumberOfOccasions}
\alias{resolveNumberOfOccasions}
\title{Resolve \code{numberOfOccasions} for a project: infer when unset, else validate.}
\usage{
resolveNumberOfOccasions(userN, modelCovariates, modelParameters)
}
\arguments{
\item{userN}{Value from \code{Evaluation}/\code{Optimization} (\code{NA} = infer).}

\item{modelCovariates}{List of covariate objects.}

\item{modelParameters}{List of \code{ModelParameter} objects.}
}
\value{
Integer number of occasions.
}
\description{
Resolve \code{numberOfOccasions} for a project: infer when unset, else validate.
}
\keyword{internal}
