% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R, R/pfim-rd-examples.R
\name{SamplingTimes}
\alias{SamplingTimes}
\title{SamplingTimes}
\usage{
SamplingTimes(outcome = character(0), samplings = numeric(0))
}
\arguments{
\item{outcome}{Character string: outcome name.}

\item{samplings}{Numeric vector: sampling times (hours or model time unit).}
}
\value{
An S7 object of class \code{SamplingTimes}.
}
\description{
Observation times for one outcome within an arm.

Times are in the model's time unit (often hours). Multiple
\code{SamplingTimes} objects on an arm cover multi-response designs.
}
\examples{
SamplingTimes(outcome = "RespPK", samplings = c(1, 4, 8))
}
