% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateTest.R
\name{saveCovariateTest}
\alias{saveCovariateTest}
\title{saveCovariateTest: save the three slots to a text file.}
\usage{
saveCovariateTest(relevanceResults, file_name, folder = NULL)
}
\arguments{
\item{relevanceResults}{A \code{CovariateTest} object.}

\item{file_name}{Output file name.}

\item{folder}{Destination folder (optional).}
}
\description{
saveCovariateTest: save the three slots to a text file.
}
