% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setEvaluationFim}
\alias{setEvaluationFim}
\title{Attach evaluated FIM results to a project}
\description{
Attach evaluated FIM results to a project

Attach evaluated FIM results to a project

Attach evaluated FIM results to a project
}
