% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{setOptimalArms}
\alias{setOptimalArms}
\title{Store optimal arm designs on the FIM object}
\description{
Store optimal arm designs on the FIM object

Store optimal arm designs on the FIM object

Store optimal arm designs on the FIM object

Store optimal arm designs on the FIM object

Store optimal arm designs on the FIM object

Store optimal arm designs on the FIM object
}
