% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{tablesForReport}
\alias{tablesForReport}
\title{FIM tables for HTML reports}
\usage{
tablesForReport(fim, evaluation, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{evaluation}{A \code{PFIMProject} object.}

\item{\dots}{Additional arguments passed to methods.}
}
\description{
FIM tables for HTML reports

FIM tables for HTML reports

FIM tables for HTML reports

FIM tables for HTML reports
}
