% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{updateSamplingTimes}
\alias{updateSamplingTimes}
\title{Refine the time grid used in response and SI plots}
\usage{
updateSamplingTimes(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{\code{samplingData} from \code{\link{getSamplingData}}.}
}
\description{
Refine the time grid used in response and SI plots
}
