% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdministrationConstraints.R,
%   R/pfim-rd-examples.R
\name{AdministrationConstraints}
\alias{AdministrationConstraints}
\title{AdministrationConstraints}
\usage{
AdministrationConstraints(outcome = character(0), doses = list())
}
\arguments{
\item{outcome}{Character string: outcome name.}

\item{doses}{List of admissible dose levels for that outcome.}
}
\value{
An S7 object of class \code{AdministrationConstraints}.
}
\description{
Allowed dose values per outcome when optimizing or reporting designs.

Used by \code{generateDosesCombination()} to build the discrete dose grid for
Fedorov-Wynn / multiplicative algorithms (and constraint reports).
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(AdministrationConstraints)
}
}
