% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates-fim.R
\name{aggregateVarianceWithCovariates}
\alias{aggregateVarianceWithCovariates}
\title{Expectation of the residual variance structure over covariate combinations and occasions.}
\usage{
aggregateVarianceWithCovariates(arm)
}
\arguments{
\item{arm}{An \code{Arm} object with nested \code{evaluationVariance}.}
}
\value{
List with \code{errorVariance} and \code{sigmaDerivatives}.
}
\description{
Expectation of the residual variance structure over covariate combinations and occasions.
}
\keyword{internal}
