% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{Arm}
\alias{Arm}
\title{Arm}
\usage{
Arm(
  name = character(0),
  size = numeric(0),
  administrations = list(),
  initialConditions = list(),
  samplingTimes = list(),
  administrationsConstraints = list(),
  samplingTimesConstraints = list(),
  evaluationModel = list(),
  evaluationGradients = list(),
  evaluationVariance = list(),
  evaluationFim = Fim(),
  covariateSampling = CovariateSamplingMethods()
)
}
\arguments{
\item{name}{Character string: arm identifier.}

\item{size}{Numeric: number of subjects in the arm.}

\item{administrations}{List of \code{Administration} objects.}

\item{initialConditions}{List of ODE initial conditions (name = variable, value = numeric).}

\item{samplingTimes}{List of \code{SamplingTimes} objects.}

\item{administrationsConstraints}{List of \code{AdministrationConstraints} objects.}

\item{samplingTimesConstraints}{List of \code{SamplingTimeConstraints} objects.}

\item{evaluationModel}{Model predictions at sampling times (filled by \code{evaluateArm}).}

\item{evaluationGradients}{Gradients of responses w.r.t. parameters (nested if covariates/IOV).}

\item{evaluationVariance}{Residual variance structure for the arm.}

\item{evaluationFim}{FIM object after \code{evaluateFim}.}

\item{covariateSampling}{Optional \code{CovariateSamplingMethods} object.}
}
\description{
One experimental arm: subjects, dosing, sampling times, and slots filled during
design evaluation (model predictions, gradients, variance, FIM).
}
