% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{armAdministration}
\alias{armAdministration}
\title{Tabular administration settings for reports}
\usage{
armAdministration(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Not used.}
}
\description{
Tabular administration settings for reports
}
