% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimeConstraints.R
\name{checkSamplingTimeConstraintsForMetaheuristic}
\alias{checkSamplingTimeConstraintsForMetaheuristic}
\title{Check sampling feasibility for PSO, PGBO or simplex}
\description{
Check sampling feasibility for PSO, PGBO or simplex
}
