% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateModelEquation.R
\name{computeCovariateValue}
\alias{computeCovariateValue}
\title{computeCovariateValue: apply beta + combinedEffect to produce the adjusted value.}
\arguments{
\item{equation}{A \code{CovariateModelEquation} subclass object.}

\item{beta}{Numeric scalar -- fixed-effect coefficient.}

\item{combinedEffect}{Numeric scalar -- combined covariate term.}
}
\value{
An updated \code{CovariateModelEquation} subclass object.
}
\description{
computeCovariateValue: apply beta + combinedEffect to produce the adjusted value.
}
