% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R, R/ModelInfusion.R
\name{convertPKModelAnalyticToPKModelODE}
\alias{convertPKModelAnalyticToPKModelODE}
\title{Convert an analytic infusion model to ODE form}
\description{
Convert an analytic infusion model to ODE form

Convert an analytic infusion model to ODE form
}
