% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateModelEquation.R
\name{CovariateModelEquation}
\alias{CovariateModelEquation}
\title{CovariateModelEquation}
\usage{
CovariateModelEquation(beta = 0, combinedEffect = 0, value = 0)
}
\arguments{
\item{beta}{Fixed-effect coefficient β (default 0).}

\item{combinedEffect}{Combined effect term Σ βₖ·covₖ (default 0).}

\item{value}{Computed adjusted value (default 0).}
}
\description{
The class \code{CovariateModelEquation} is the base class for the two
covariate structural equations supported by PFIM:
\code{Additive} and \code{Exponential}.

A covariate model equation adjusts the population mean for individual i
with covariate value
\itemize{
  \item \code{Additive}:    additive model
  \item \code{Exponential}: exponential model
}
}
