% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateTest.R
\name{CovariateTest-class}
\alias{CovariateTest-class}
\alias{CovariateTest}
\title{CovariateTest Class}
\arguments{
\item{covariate}{\code{data.frame} for \eqn{\beta} significance.}

\item{parameter}{Deprecated; kept empty (typical \eqn{\mu} are not reported).}

\item{nonRelevance}{\code{data.frame} for TOST non-relevance on \eqn{\beta}.}

\item{relevance}{\code{data.frame} for clinical relevance on \eqn{\beta}.}

\item{settings}{List of run/display options.}
}
\value{
An S7 object of class \code{CovariateTest}.
}
\description{
\code{CovariateTest} stores covariate-test results.
}
\details{
\describe{
  \item{covariate}{Significance on covariate effects \eqn{\beta}}
  \item{nonRelevance}{TOST non-relevance on \eqn{\beta}}
  \item{relevance}{Clinical relevance on \eqn{\beta}}
  \item{settings}{Run options (\code{tests}, \code{thetaL/U}, \code{target_power}, \code{alpha}, \code{N0})}
}
}
\examples{
\donttest{
source(system.file("examples", "covariate-test-minimal.R", package = "PFIM"))
class(ct)
}

}
