% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{defineModelAdministration}
\alias{defineModelAdministration}
\title{Bind dosing schedules and sampling grids to the model}
\usage{
defineModelAdministration(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{\code{arm} for model methods.}
}
\description{
Bind dosing schedules and sampling grids to the model
}
