% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{definePKModel}
\alias{definePKModel}
\title{Remap library PK equations onto project compartments}
\usage{
definePKModel(pkModel, pfimproject, ...)
}
\arguments{
\item{pkModel}{Library PK model object.}

\item{pfimproject}{A \code{PFIMProject}, \code{Evaluation}, or \code{Optimization} object.}

\item{...}{Optional method arguments.}
}
\description{
Remap library PK equations onto project compartments
}
