% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{definePKPDModel}
\alias{definePKPDModel}
\title{Combine PK and PD library equations for the project}
\usage{
definePKPDModel(pkModel, pdModel, pfimproject, ...)
}
\arguments{
\item{pkModel}{Library PK model object.}

\item{pdModel}{Library PD model object.}

\item{pfimproject}{A \code{PFIMProject}, \code{Evaluation}, or \code{Optimization} object.}

\item{...}{Optional method arguments.}
}
\description{
Combine PK and PD library equations for the project
}
