% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalytic.R
\name{evaluateAnalyticCore}
\alias{evaluateAnalyticCore}
\title{evaluateAnalyticCore: core analytic evaluation (shared by covariate path).}
\usage{
evaluateAnalyticCore(model, arm)
}
\arguments{
\item{model}{A \code{ModelAnalytic} object.}

\item{arm}{An \code{Arm} object.}
}
\description{
evaluateAnalyticCore: core analytic evaluation (shared by covariate path).
}
\keyword{internal}
