% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{evaluateArm}
\alias{evaluateArm}
\title{Evaluate model, gradients, variance and FIM for one arm}
\usage{
evaluateArm(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{\code{model} and \code{fim} for the \code{Arm} method.}
}
\description{
Evaluate model, gradients, variance and FIM for one arm
}
