% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{evaluateInitialConditions}
\alias{evaluateInitialConditions}
\title{Evaluate ODE initial conditions from the arm}
\usage{
evaluateInitialConditions(model, ...)
}
\arguments{
\item{model}{A \code{Model} object.}

\item{...}{\code{arm} and optionally \code{doseEvent} for bolus ODE models.}
}
\description{
Evaluate ODE initial conditions from the arm
}
