% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedorovWynnAlgorithm.R
\name{FedorovWynnAlgorithm_Rcpp}
\alias{FedorovWynnAlgorithm_Rcpp}
\title{Fedorov-Wynn algorithm in Rcpp.
Wrapper around the compiled routine in \code{src/Fedorovwynnalgorithm.cpp}}
\arguments{
\item{protocols}{List of elementary protocols.}

\item{ndimen}{Integer vector of dimensions.}

\item{nbprot}{Integer vector of protocol counts.}

\item{numprot}{Integer vector of protocol indices.}

\item{freq}{Numeric vector of frequencies.}

\item{nbdata}{Integer vector of data counts.}

\item{vectps}{Numeric vector of sampling times.}

\item{fisher}{Numeric vector of Fisher information values.}

\item{error}{Integer error/status code.}

\item{protdep}{Integer protocol-dependence flags.}

\item{freqdep}{Numeric frequency-dependence values.}
}
\value{
A list with the results of the Fedorov-Wynn algorithm.
}
\description{
Fedorov-Wynn algorithm in Rcpp.
Wrapper around the compiled routine in \code{src/Fedorovwynnalgorithm.cpp}
}
\keyword{internal}
