% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{generateReportEvaluation}
\alias{generateReportEvaluation}
\title{Render the evaluation HTML report}
\usage{
generateReportEvaluation(fim, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{\dots}{Additional arguments passed to methods.}
}
\description{
Render the evaluation HTML report

Render the evaluation HTML report

Render the evaluation HTML report

Render the evaluation HTML report
}
