% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R
\name{generateReportOptimization}
\alias{generateReportOptimization}
\title{Render the optimization HTML report}
\usage{
generateReportOptimization(fim, optimizationAlgorithm, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{optimizationAlgorithm}{A \code{Optimization} object.}

\item{\dots}{Additional arguments passed to methods.}
}
\description{
Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report

Render the optimization HTML report
}
