% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getArmConstraints}
\alias{getArmConstraints}
\title{Administration and sampling constraints for an optimizer}
\usage{
getArmConstraints(arm, optimizationAlgorithm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{optimizationAlgorithm}{An optimization algorithm object.}

\item{...}{Not used.}
}
\description{
Administration and sampling constraints for an optimizer
}
