% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R, R/pfim-rd-examples.R
\name{getArmData}
\alias{getArmData}
\title{Dose and sampling summary for reports}
\usage{
getArmData(arm, ...)
}
\arguments{
\item{arm}{First argument of generic.}

\item{...}{Not used.}
}
\value{
List of arm-level dose and sampling summaries by outcome.
}
\description{
Dose and sampling summary for reports

Build dose / sampling summary rows for HTML and console reports.
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getArmData)
}
}
\keyword{internal}
