% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R, R/Optimization.R,
%   R/evaluation-accessors.R, R/pfim-rd-examples.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Retrieve correlation matrix from results}
\usage{
getCorrelationMatrix(pfimproject, ...)
}
\arguments{
\item{pfimproject}{A \code{PFIMProject} object.}

\item{...}{Optional method arguments.}
}
\value{
Correlation matrix of the parameter estimates.
}
\description{
Retrieve correlation matrix from results

Parameter correlations from the optimal design FIM.

Parameter correlations from the evaluated design FIM.
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(getCorrelationMatrix)
}
}
