% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalCovariate.R,
%   R/CategoricalCovariateWithIOV.R
\name{getCovariateEffects}
\alias{getCovariateEffects}
\title{Estimated covariate effects on parameters}
\description{
Estimated covariate effects on parameters

Estimated covariate effects on parameters
}
