% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/evaluation-accessors.R
\name{getDeterminant}
\alias{getDeterminant}
\title{Determinant of the optimal design FIM}
\description{
Determinant of the optimal design FIM

Determinant of the optimal design FIM
}
