% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/evaluation-accessors.R
\name{getFisherMatrix}
\alias{getFisherMatrix}
\title{Fisher matrix blocks for the optimal design}
\description{
Fisher matrix blocks for the optimal design

Fisher matrix blocks for the optimal design
}
