% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{getNumberOfOccasionsForModel}
\alias{getNumberOfOccasionsForModel}
\title{Occasion count for a built \code{Model} object.}
\usage{
getNumberOfOccasionsForModel(model)
}
\arguments{
\item{model}{A \code{Model} object.}
}
\value{
Integer number of occasions.
}
\description{
Occasion count for a built \code{Model} object.
}
\keyword{internal}
