% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{getSamplingData}
\alias{getSamplingData}
\title{Sampling times and grid extent for evaluation plots}
\usage{
getSamplingData(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Not used.}
}
\description{
Sampling times and grid extent for evaluation plots
}
