% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPDModels.R
\name{LibraryOfPDModels}
\alias{LibraryOfPDModels}
\title{LibraryOfPDModels}
\usage{
LibraryOfPDModels(models = list())
}
\arguments{
\item{models}{Named list of built-in PD model equation definitions.}
}
\description{
Built-in pharmacodynamic (PD) model library (Emax, turnover, etc.).
}
