% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R
\name{LibraryOfPKModels}
\alias{LibraryOfPKModels}
\title{LibraryOfPKModels}
\usage{
LibraryOfPKModels(models = list())
}
\arguments{
\item{models}{Named list of built-in PK model equation definitions.}
}
\description{
Built-in pharmacokinetic (PK) model library (bolus, infusion, absorption, 1–3 compartments).
}
