% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGBOAlgorithm.R, R/PSOAlgorithm.R,
%   R/SimplexAlgorithm.R, R/FedorovWynnAlgorithm.R, R/MultiplicativeAlgorithm.R
\name{optimizeDesign}
\alias{optimizeDesign}
\title{Search for a D-optimal sampling design}
\description{
Search for a D-optimal sampling design

Search for a D-optimal sampling design

Search for a D-optimal sampling design

Search for a D-optimal sampling design

Search for a D-optimal sampling design
}
