% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPDModels.R
\docType{data}
\name{pdModelLibrary}
\alias{pdModelLibrary}
\title{Default built-in PD model library instance.}
\format{
An S7 object of class \code{LibraryOfPDModels}.
}
\usage{
pdModelLibrary
}
\description{
Default built-in PD model library instance.
}
\seealso{
\code{\link{LibraryOfPDModels}} for the class constructor.
}
\keyword{datasets}
