% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim-session.R
\name{pfim_reset_session}
\alias{pfim_reset_session}
\title{Reset PFIM session options to package defaults}
\usage{
pfim_reset_session()
}
\value{
Invisibly \code{NULL}.
}
\description{
Reset PFIM session options to package defaults
}
