% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryOfPKModels.R
\name{pkModelLibrary}
\alias{pkModelLibrary}
\title{Default built-in PK model library instance.}
\format{
An S7 object of class \code{LibraryOfPKModels}.
}
\usage{
pkModelLibrary
}
\value{
A \code{LibraryOfPKModels} object.
}
\description{
Default built-in PK model library instance.
}
\examples{
\donttest{
length(prop(pkModelLibrary, "models"))
}
}
\seealso{
\code{\link{LibraryOfPKModels}} for the class constructor.
}
