% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-accessors.R, R/evaluation-plots.R
\name{plotEvaluation}
\alias{plotEvaluation}
\title{Response plots for all designs in an evaluation}
\description{
Response plots for all designs in an evaluation

Response plots for all designs in an evaluation
}
