% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationResults}
\alias{plotEvaluationResults}
\title{ggplot of predicted responses with sampling points}
\usage{
plotEvaluationResults(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Evaluation outputs and plot options for the \code{Arm} method.}
}
\description{
ggplot of predicted responses with sampling points
}
