% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{plotEvaluationSI}
\alias{plotEvaluationSI}
\title{ggplot of parameter sensitivities over time}
\usage{
plotEvaluationSI(arm, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{...}{Method-specific sensitivity-index plotting inputs (see methods).}
}
\value{
Nested list of ggplot objects for sensitivity indices.
}
\description{
ggplot of parameter sensitivities over time

Default method for \code{Arm}.
}
\keyword{internal}
