% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/IndividualFim.R, R/PopulationFim.R
\name{plotRSEFIM}
\alias{plotRSEFIM}
\title{Barplot of relative standard errors from a FIM}
\usage{
plotRSEFIM(fim, evaluation, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{evaluation}{A \code{PFIMProject} object.}

\item{\dots}{Additional arguments passed to methods.}
}
\description{
Barplot of relative standard errors from a FIM

Barplot of relative standard errors from a FIM

Barplot of relative standard errors from a FIM
}
