% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R, R/evaluation-plots.R
\name{plotSensitivityIndices}
\alias{plotSensitivityIndices}
\title{Sensitivity indices for the optimal design}
\description{
Sensitivity indices for the optimal design

Sensitivity indices for the optimal design
}
