% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R
\name{plotShrinkage}
\alias{plotShrinkage}
\title{plotShrinkage: plot Bayesian shrinkage bars.}
\usage{
plotShrinkage(fim, evaluation, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{evaluation}{A \code{PFIMProject} object.}

\item{\dots}{Additional arguments passed to methods.}
}
\description{
plotShrinkage: plot Bayesian shrinkage bars.
}
