% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicativeAlgorithm.R,
%   R/pfim-rd-examples.R
\name{plotWeightsMultiplicativeAlgorithm}
\alias{plotWeightsMultiplicativeAlgorithm}
\title{Weight trajectories of the multiplicative algorithm}
\value{
A \code{ggplot2} plot object.
}
\description{
Weight trajectories of the multiplicative algorithm
}
\examples{
\dontrun{
# help(plotWeightsMultiplicativeAlgorithm)
}
}
\keyword{internal}
