% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{processArmEvaluationResults}
\alias{processArmEvaluationResults}
\title{Model response plots for one arm}
\usage{
processArmEvaluationResults(arm, model, fim, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{model}{A \code{Model} object.}

\item{fim}{A \code{Fim} object.}

\item{...}{Design label and plot options (see methods).}
}
\value{
Nested list of response plots by design and arm.
}
\description{
Re-evaluates the model on a dense time grid and overlays the
design sampling times as red markers. The FIM arm is not modified.
}
\keyword{internal}
