% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arm.R
\name{processArmEvaluationSI}
\alias{processArmEvaluationSI}
\title{Sensitivity-index plots for one arm}
\usage{
processArmEvaluationSI(arm, model, fim, ...)
}
\arguments{
\item{arm}{An \code{Arm} object.}

\item{model}{A \code{Model} object.}

\item{fim}{A \code{Fim} object.}

\item{...}{\code{designName} and \code{plotOptions}.}
}
\description{
Sensitivity-index plots for one arm
}
