% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportional.R
\name{Proportional}
\alias{Proportional}
\title{Proportional}
\usage{
Proportional(
  output = character(0),
  equation = expression(sigmaSlope),
  derivatives = list(),
  sigmaInter = 0,
  sigmaSlope = 0,
  sigmaInterFixed = FALSE,
  sigmaSlopeFixed = FALSE,
  cError = 1
)
}
\arguments{
\item{output}{Character string: outcome name.}

\item{equation}{Expression defining residual variance as a function of predictions.}

\item{derivatives}{List of partial derivatives w.r.t. error parameters.}

\item{sigmaInter}{Additive error standard deviation (inter-subject / residual).}

\item{sigmaSlope}{Proportional error standard deviation (slope).}

\item{sigmaInterFixed}{If TRUE, \code{sigmaInter} is fixed in the FIM.}

\item{sigmaSlopeFixed}{If TRUE, \code{sigmaSlope} is fixed in the FIM.}

\item{cError}{Power parameter in the error model (when applicable).}
}
\description{
Proportional residual error: variance proportional to the prediction.
}
