% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOAlgorithm.R
\name{PSOAlgorithm}
\alias{PSOAlgorithm}
\title{PSOAlgorithm}
\usage{
PSOAlgorithm(
  maxIteration = numeric(0),
  populationSize = numeric(0),
  seed = numeric(0),
  personalLearningCoefficient = numeric(0),
  globalLearningCoefficient = numeric(0),
  showProcess = FALSE
)
}
\arguments{
\item{maxIteration}{Numeric: the maxIteration.}

\item{populationSize}{Numeric: the populationSize.}

\item{seed}{Numeric: the seed.}

\item{personalLearningCoefficient}{Numeric: the personalLearningCoefficient.}

\item{globalLearningCoefficient}{Numeric: the globalLearningCoefficient.}

\item{showProcess}{Logical: the showProcess.}
}
\description{
Particle Swarm Optimization for sampling-time search (Rcpp kernel in \code{src/Psokernel.cpp}).
}
