% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingTimes.R
\name{SamplingTimes}
\alias{SamplingTimes}
\title{SamplingTimes}
\usage{
SamplingTimes(outcome = character(0), samplings = numeric(0))
}
\arguments{
\item{outcome}{Character string: outcome name.}

\item{samplings}{Numeric vector: sampling times (hours or model time unit).}
}
\description{
Observation times for one outcome within an arm.
}
