% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/pfim-rd-examples.R
\name{setOptimalArms}
\alias{setOptimalArms}
\title{Optimal arms from design optimization}
\usage{
setOptimalArms(fim, optimizationAlgorithm, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{optimizationAlgorithm}{Optimizer object from \code{run(Optimization)}.}

\item{...}{Not used by current methods.}
}
\value{
The modified \code{Fim} object.
}
\description{
Optimal arms from design optimization
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(setOptimalArms)
}
}
\keyword{internal}
