% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFIMProject.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show}
\alias{show,PFIM::Optimization-method}
\alias{show,PFIM::Evaluation-method}
\alias{show,PFIM::CovariateTest-method}
\title{Show methods for PFIM objects}
\usage{
show(object)
}
\arguments{
\item{object}{A PFIM object (\code{Evaluation}, \code{Optimization},
\code{CovariateTest}, etc.).}
}
\description{
Console display for PFIM S7 classes. Dispatches via \code{methods::show}.
}
\section{Methods}{

\describe{
  \item{\code{Optimization}}{Display initial and optimal design results.}
  \item{\code{Evaluation}}{Display evaluation results (FIM, SE, RSE, etc.).}
  \item{\code{CovariateTest}}{Display covariate-test results.}
}
}

\keyword{methods}
