% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexAlgorithm.R
\name{SimplexAlgorithm}
\alias{SimplexAlgorithm}
\title{SimplexAlgorithm}
\usage{
SimplexAlgorithm(
  pctInitialSimplexBuilding = numeric(0),
  maxIteration = numeric(0),
  seed = numeric(0),
  tolerance = numeric(0),
  showProcess = FALSE
)
}
\arguments{
\item{pctInitialSimplexBuilding}{Numeric: the pctInitialSimplexBuilding.}

\item{maxIteration}{Numeric: the maxIteration.}

\item{seed}{Numeric: the seed.}

\item{tolerance}{Numeric: the tolerance.}

\item{showProcess}{Logical: the showProcess.}
}
\description{
Nelder-Mead (amoeba) optimization for sampling times (Rcpp kernel).
}
