% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fim.R, R/BayesianFim.R, R/IndividualFim.R,
%   R/PopulationFim.R, R/pfim-rd-examples.R
\name{tablesForReport}
\alias{tablesForReport}
\title{FIM tables for HTML reports}
\usage{
tablesForReport(fim, evaluation, ...)
}
\arguments{
\item{fim}{A \code{Fim} object.}

\item{evaluation}{A \code{PFIMProject} evaluation object.}

\item{...}{Not used by current methods.}
}
\value{
Report tables as a list or data frame.
}
\description{
FIM tables for HTML reports

FIM tables for HTML reports

FIM tables for HTML reports

FIM tables for HTML reports
}
\examples{
\donttest{
source(system.file("examples", "evaluation-minimal.R", package = "PFIM"))
# help(tablesForReport)
}
}
\keyword{internal}
